KAS-1402: remove useage of manage-minister mixin
