+++ app/components/utils/edit-minister-modal/template.hbs
@@ -9,7 +9,6 @@
         singleSelect=true
         readOnly=true
         selectedMandatees=selectedMandatee
-        chooseMandatee=(action "mandateeSelected")
       }}
     </div>
     {{#each rowToShow.domains as |domain|}}
